officially need to take a break
brain is fried and i have got myself into a pickle where the mutexes needed to update chunk geometry are also needed to be locked to draw the chunk geometry, so the chunk is flickering when updated. Also, something in the rebuildqueue is queueing more stuff so it is causing loops of rebuilds.